+++ packages/integration-sdk-runtime/src/synchronization/index.ts
@@ -21,10 +21,10 @@ import { FlushedGraphObjectData } from '../storage/types';
 import { AttemptContext, retry } from '@lifeomic/attempt';
 import { randomUUID as uuid } from 'crypto';
 import { createEventPublishingQueue } from './events';
-import { AxiosInstance } from 'axios';
 import { iterateParsedGraphFiles } from '..';
 import { shrinkBatchRawData } from './shrinkBatchRawData';
 import { batchGraphObjectsBySizeInBytes, getSizeOfObject } from './batchBySize';
+import type { Alpha } from '@lifeomic/alpha';
 
 export { synchronizationApiError };
 export { createEventPublishingQueue } from './events';
@@ -351,7 +351,7 @@ export interface UploadDataLookup {
 
 interface UploadDataChunkParams<T extends UploadDataLookup, K extends keyof T> {
   logger: IntegrationLogger;
-  apiClient: AxiosInstance;
+  apiClient: Alpha;
   jobId: string;
   type: K;
   batch: T[K][];